// src/types/gen/mikan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type MikanResourceItem = {
  title: string;
  page_url: string;
  torrent_url?: string;
  magnet?: string;
  pub_date?: string;
  size_bytes?: bigint;
  group?: string;
  episode?: number;
  episode_range?: string;
  resolution?: number;
  subtitle_lang?: string;
  subtitle_type?: string;
};

export type MikanResourcesResponse = {
  mapped: boolean;
  mikan_bangumi_id?: number;
  items: Array<MikanResourceItem>;
};